+++ src/Commands/UiCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace HardImpact\Orbit\Ui\Commands;
+namespace HardImpact\Orbit\App\Commands;
 
 use Illuminate\Console\Command;