+++ src/Http/Controllers/EnvironmentController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace HardImpact\Orbit\Ui\Http\Controllers;
+namespace HardImpact\Orbit\App\Http\Controllers;
 
 use HardImpact\Orbit\Core\Models\Environment;
 use HardImpact\Orbit\Core\Models\Setting;
@@ -12,7 +12,7 @@ use HardImpact\Orbit\Core\Services\MacPhpFpmConfigService;
 use HardImpact\Orbit\Core\Services\NotificationService;
 use HardImpact\Orbit\Core\Services\OrbitCli\ConfigurationService;
 use HardImpact\Orbit\Core\Services\OrbitCli\StatusService;
-use HardImpact\Orbit\Ui\Http\Controllers\Concerns\ProvidesRemoteApiUrl;
+use HardImpact\Orbit\App\Http\Controllers\Concerns\ProvidesRemoteApiUrl;
 use Illuminate\Http\Request;
 
 /**