+++ tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use HardImpact\Orbit\Ui\Tests\TestCase;
+use HardImpact\Orbit\App\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);