+++ tests/TestCase.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace HardImpact\Orbit\Ui\Tests;
+namespace HardImpact\Orbit\App\Tests;
 
-use HardImpact\Orbit\Ui\OrbitAppServiceProvider;
+use HardImpact\Orbit\App\OrbitAppServiceProvider;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Inertia\ServiceProvider as InertiaServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;